requirements.txt
pandas>=2.0.0
numpy>=1.26.0
python-dotenv>=1.0.0
requests>=2.31.0
yfinance>=0.2.40
scipy>=1.12.0
